chore: clean up hooks
