+++ src/hooks/useCanvasLayout.ts
@@ -1,0 +1,20 @@
+import { useCallback, useState } from "react";
+
+import type { EngineState } from "../useLivelineEngine";
+import type { LayoutChangeEvent } from "react-native";
+
+export function useCanvasLayout(engine: EngineState) {
+  const [layoutHeight, setLayoutHeight] = useState(0);
+
+  const onLayout = useCallback(
+    (e: LayoutChangeEvent) => {
+      const { width, height } = e.nativeEvent.layout;
+      engine.canvasWidth.value = width;
+      engine.canvasHeight.value = height;
+      setLayoutHeight(height);
+    },
+    [engine.canvasWidth, engine.canvasHeight],
+  );
+
+  return { layoutHeight, onLayout } as const;
+}
